+++ Expanded Cheat Menu/Code/GedSocket.lua
@@ -27,7 +27,7 @@ function GedSocket:Done()
 		end
 	end
 end
-function GedSocket:OnDisconnect()
+function GedSocket:OnDisconnect(reason)
 	self:Done()
 end
 function GedSocket:rpcGedQuit()
@@ -54,7 +54,7 @@ function GedSocket:UnbindObj(name, to_prefix)
 	if to_prefix then
 		local pref = name .. to_prefix
 		for obj_name in pairs(self.bound_objects) do
-			if obj_name:starts_with(pref) then
+			if string.starts_with(obj_name, pref) then
 				self.bound_objects[obj_name] = nil
 			end
 		end