+++ Expanded Cheat Menu/Code/ReplacedFunctions.lua
@@ -16,14 +16,24 @@ local SetDlgTrans = ChoGGi.ComFuncs.SetDlgTrans
 local RetName = ChoGGi.ComFuncs.RetName
 local IsValidXWin = ChoGGi.ComFuncs.IsValidXWin
 local GetParentOfKind = ChoGGi.ComFuncs.GetParentOfKind
+local DotPathToObject = ChoGGi.ComFuncs.DotPathToObject
 
 local UserSettings = ChoGGi.UserSettings
 local testing = ChoGGi.testing
+
+local g_env = _G
+local function AddToOrigs(name)
+	if not ChoGGi.OrigFuncs[name] then
+		ChoGGi.OrigFuncs[name] = DotPathToObject(name)
+	end
+end
+
 
 do -- non-class obj funcs
 
 	-- this will reset override, so we sleep and reset it
 	local ChoOrig_ClosePlanetCamera = ClosePlanetCamera
+	AddToOrigs("ClosePlanetCamera")
 	function ClosePlanetCamera(...)
 		if UserSettings.Lightmodel then
 			CreateRealTimeThread(function()
@@ -36,6 +46,7 @@ do -- non-class obj funcs
 
 	-- don't trigger quakes if setting is enabled
 	local ChoOrig_TriggerMarsquake = TriggerMarsquake
+	AddToOrigs("TriggerMarsquake")
 	function TriggerMarsquake(...)
 		if not UserSettings.DisasterQuakeDisable then
 			return ChoOrig_TriggerMarsquake(...)
@@ -44,6 +55,7 @@ do -- non-class obj funcs
 
 	-- don't trigger toxic rains if setting is enabled
 	local ChoOrig_RainProcedure = RainProcedure
+	AddToOrigs("RainProcedure")
 	function RainProcedure(settings, ...)
 		if settings.type == "normal" or not UserSettings.DisasterRainsDisable then
 			return ChoOrig_RainProcedure(settings, ...)
@@ -57,6 +69,7 @@ do -- non-class obj funcs
 
 	-- stops mod editor itself
 	local ChoOrig_OpenGedApp = OpenGedApp
+	AddToOrigs("OpenGedApp")
 	function OpenGedApp(template, ...)
 		if template == "ModsEditor" and (testing or UserSettings.SkipModEditorDialog) then
 			return
@@ -66,6 +79,7 @@ do -- non-class obj funcs
 
 	-- get rid of "This savegame was loaded in the past without required mods or with an incompatible game version."
 	local ChoOrig_WaitMarsMessage = WaitMarsMessage
+	AddToOrigs("WaitMarsMessage")
 	function WaitMarsMessage(parent, title, msg, ...)
 		if (testing or UserSettings.SkipIncompatibleModsMsg) and IsT(msg) == 10888 then
 			return
@@ -75,6 +89,7 @@ do -- non-class obj funcs
 
 	-- examine persist errors (if any)
 	local ChoOrig_ReportPersistErrors = ReportPersistErrors
+	AddToOrigs("ReportPersistErrors")
 	function ReportPersistErrors(...)
 		local errors, warnings = 0, 0
 		if UserSettings.DebugPersistSaves and __error_table__ and #__error_table__ > 0 then
@@ -87,6 +102,7 @@ do -- non-class obj funcs
 	end
 
 	local ChoOrig_TGetID = TGetID
+	AddToOrigs("TGetID")
 	function TGetID(t, ...)
 		local t_type = type(t)
 		if t_type ~= "table" and t_type ~= "userdata" then
@@ -97,6 +113,7 @@ do -- non-class obj funcs
 
 	-- I guess, don't pass a string to it?
 	local ChoOrig_TDevModeGetEnglishText = TDevModeGetEnglishText
+	AddToOrigs("TDevModeGetEnglishText")
 	function TDevModeGetEnglishText(t, ...)
 		if type(t) == "string" then
 			return t
@@ -106,6 +123,7 @@ do -- non-class obj funcs
 
 	-- fix for sending nil id to it
 	local ChoOrig_LoadCustomOnScreenNotification = LoadCustomOnScreenNotification
+	AddToOrigs("LoadCustomOnScreenNotification")
 	function LoadCustomOnScreenNotification(notification, ...)
 		-- the first return is id, and some mods (cough Ambassadors cough) send a nil id, which breaks the func
 		if table.unpack(notification) then
@@ -115,6 +133,7 @@ do -- non-class obj funcs
 
 	-- change rocket cargo cap
 	local ChoOrig_GetMaxCargoShuttleCapacity = GetMaxCargoShuttleCapacity
+	AddToOrigs("GetMaxCargoShuttleCapacity")
 	function GetMaxCargoShuttleCapacity(...)
 		return UserSettings.StorageShuttle or ChoOrig_GetMaxCargoShuttleCapacity(...)
 	end
@@ -122,6 +141,7 @@ do -- non-class obj funcs
 
 	-- report building as not-a-wonder to the func that checks for wonders
 	local ChoOrig_UIGetBuildingPrerequisites = UIGetBuildingPrerequisites
+	AddToOrigs("UIGetBuildingPrerequisites")
 	function UIGetBuildingPrerequisites(cat_id, template, ...)
 		-- missing dlc
 		if BuildingTemplates[template.id] then
@@ -146,6 +166,7 @@ do -- non-class obj funcs
 
 	-- stops confirmation dialog about missing mods (still lets you know they're missing)
 	local ChoOrig_GetMissingMods = GetMissingMods
+	AddToOrigs("GetMissingMods")
 	function GetMissingMods(...)
 		if UserSettings.SkipMissingMods then
 			return "", false
@@ -165,6 +186,7 @@ do -- non-class obj funcs
 			local name = dlc_funcs[i]
 
 			local ChoOrig_func = _G[name]
+			AddToOrigs(name)
 			_G[name] = function(dlc, ...)
 				-- stuff added for future dlc is showing up and erroring out
 				if not dlc or dlc == "" then
@@ -178,6 +200,7 @@ do -- non-class obj funcs
 
 	-- lets you load saved games that have dlc
 	local ChoOrig_IsDlcRequired = IsDlcRequired
+	AddToOrigs("IsDlcRequired")
 	function IsDlcRequired(dlc, ...)
 		-- stuff added for future dlc is showing up and erroring out
 		if not dlc or dlc == "" then
@@ -189,6 +212,7 @@ do -- non-class obj funcs
 
 	-- always able to show console
 	local ChoOrig_ShowConsole = ShowConsole
+	AddToOrigs("ShowConsole")
 	function ShowConsole(visible, ...)
 		if visible then
 			-- ShowConsole checks for this
@@ -199,6 +223,7 @@ do -- non-class obj funcs
 
 	-- console stuff
 	local ChoOrig_ShowConsoleLog = ShowConsoleLog
+	AddToOrigs("ShowConsoleLog")
 	function ShowConsoleLog(visible, ...)
 		-- we only want to show it if it's enabled or we're in mod editor mode
 		visible = UserSettings.ConsoleToggleHistory or ChoGGi.ComFuncs.ModEditorActive()
@@ -223,6 +248,7 @@ do -- non-class obj funcs
 		end
 
 		local ChoOrig_ShowPopupNotification = ShowPopupNotification
+		AddToOrigs("ShowPopupNotification")
 		function ShowPopupNotification(preset, ...)
 			if UserSettings.DisableHints and suggestions[preset] then
 				return
@@ -233,12 +259,14 @@ do -- non-class obj funcs
 
 	-- UI transparency dialogs (buildmenu, pins, infopanel)
 	local ChoOrig_OpenDialog = OpenDialog
+	AddToOrigs("OpenDialog")
 	function OpenDialog(...)
 		return SetDlgTrans(ChoOrig_OpenDialog(...))
 	end
 
 	-- skips story bit dialogs
 	local ChoOrig_PopupNotificationBegin = PopupNotificationBegin
+	AddToOrigs("PopupNotificationBegin")
 	function PopupNotificationBegin(dlg, ...)
 		if UserSettings.SkipStoryBitsDialogs and dlg.context and dlg.context.is_storybit
 		then
@@ -263,37 +291,44 @@ do -- func exists before classes
 
 	-- update production (OnMsgs.lua)
 	local ChoOrig_SingleResourceProducer_Init = SingleResourceProducer.Init
+	AddToOrigs("SingleResourceProducer.Init")
 	function SingleResourceProducer:Init(...)
 		ChoOrig_SingleResourceProducer_Init(self, ...)
 		Msg("ChoGGi_SpawnedProducer", self, "production_per_day")
 	end
 	local ChoOrig_AirProducer_CreateLifeSupportElements = AirProducer.CreateLifeSupportElements
+	AddToOrigs("AirProducer.CreateLifeSupportElements")
 	function AirProducer:CreateLifeSupportElements(...)
 		ChoOrig_AirProducer_CreateLifeSupportElements(self, ...)
 		Msg("ChoGGi_SpawnedProducer", self, "air_production")
 	end
 	local ChoOrig_WaterProducer_CreateLifeSupportElements = WaterProducer.CreateLifeSupportElements
+	AddToOrigs("WaterProducer.CreateLifeSupportElements")
 	function WaterProducer:CreateLifeSupportElements(...)
 		ChoOrig_WaterProducer_CreateLifeSupportElements(self, ...)
 		Msg("ChoGGi_SpawnedProducer", self, "water_production")
 	end
 	local ChoOrig_ElectricityProducer_CreateElectricityElement = ElectricityProducer.CreateElectricityElement
+	AddToOrigs("ElectricityProducer.CreateElectricityElement")
 	function ElectricityProducer:CreateElectricityElement(...)
 		ChoOrig_ElectricityProducer_CreateElectricityElement(self, ...)
 		Msg("ChoGGi_SpawnedProducer", self, "electricity_production")
 	end
 	local ChoOrig_PinnableObject_TogglePin = PinnableObject.TogglePin
+	AddToOrigs("PinnableObject.TogglePin")
 	function PinnableObject:TogglePin(...)
 		ChoOrig_PinnableObject_TogglePin(self, ...)
 		Msg("ChoGGi_TogglePinnableObject", self)
 	end
 	local ChoOrig_Drone_GameInit = Drone.GameInit
+	AddToOrigs("Drone.GameInit")
 	function Drone:GameInit(...)
 		ChoOrig_Drone_GameInit(self, ...)
 		-- slight delay
 		CreateRealTimeThread(Msg, "ChoGGi_SpawnedDrone", self)
 	end
 	local ChoOrig_BaseBuilding_GameInit = BaseBuilding.GameInit
+	AddToOrigs("BaseBuilding.GameInit")
 	function BaseBuilding:GameInit(...)
 		ChoOrig_BaseBuilding_GameInit(self, ...)
 		-- slight delay
@@ -313,6 +348,7 @@ function OnMsg.ClassesGenerate()
 
 	-- needed for SetDesiredAmount in depots
 	local ChoOrig_ResourceStockpileBase_GetMax = ResourceStockpileBase.GetMax
+	AddToOrigs("ResourceStockpileBase.GetMax")
 	function ResourceStockpileBase:GetMax(...)
 		if UserSettings.StorageUniversalDepot and self.template_name == "UniversalStorageDepot" then
 			return UserSettings.StorageUniversalDepot / const.ResourceScale
@@ -325,6 +361,7 @@ function OnMsg.ClassesGenerate()
 
 		-- no more limit to R+T
 		local ChoOrig_LandscapeConstructionController_Activate = LandscapeConstructionController.Activate
+		AddToOrigs("LandscapeConstructionController.Activate")
 		function LandscapeConstructionController:Activate(...)
 			if UserSettings.RemoveLandScapingLimits then
 				self.brush_radius_step = 100
@@ -340,6 +377,7 @@ function OnMsg.ClassesGenerate()
 
 		-- stop drones/rovers from slowing down in dustdevils
 		local ChoOrig_DroneBase_RegisterDustDevil = DroneBase.RegisterDustDevil
+		AddToOrigs("DroneBase.RegisterDustDevil")
 		function DroneBase:RegisterDustDevil(devil, ...)
 			if (UserSettings.SpeedWaspDrone and self:IsKindOf("FlyingDrone"))
 				or (UserSettings.SpeedDrone and self:IsKindOf("Drone") and not self:IsKindOf("FlyingDrone"))
@@ -353,6 +391,7 @@ function OnMsg.ClassesGenerate()
 
 	-- all storybit/neg/etc options enabled
 	local ChoOrig_Condition_Evaluate = Condition.Evaluate
+	AddToOrigs("Condition.Evaluate")
 	function Condition.Evaluate(...)
 		return UserSettings.OverrideConditionPrereqs
 			or ChoOrig_Condition_Evaluate(...)
@@ -367,6 +406,7 @@ function OnMsg.ClassesGenerate()
 		end
 
 		local ChoOrig_InfopanelItems_Open = InfopanelItems.Open
+		AddToOrigs("InfopanelItems.Open")
 		function InfopanelItems:Open(...)
 			if UserSettings.LimitCropsUIWidth then
 				self:SetMaxWidth(width - Dialogs.Infopanel.box:sizex())
@@ -397,11 +437,13 @@ function OnMsg.ClassesGenerate()
 		end
 
 		local ChoOrig_MechanizedDepot_CheatFill = MechanizedDepot.CheatFill
+		AddToOrigs("MechanizedDepot.CheatFill")
 		function MechanizedDepot.CheatFill(...)
 			return SuspendAndFire(ChoOrig_MechanizedDepot_CheatFill, ...)
 		end
 
 		local ChoOrig_UniversalStorageDepot_CheatFill = UniversalStorageDepot.CheatFill
+		AddToOrigs("UniversalStorageDepot.CheatFill")
 		function UniversalStorageDepot.CheatFill(...)
 			return SuspendAndFire(ChoOrig_UniversalStorageDepot_CheatFill, ...)
 		end
@@ -409,7 +451,9 @@ function OnMsg.ClassesGenerate()
 
 	-- that's what we call a small font
 	local ChoOrig_XWindow_UpdateMeasure = XWindow.UpdateMeasure
+	AddToOrigs("XWindow.UpdateMeasure")
 	local ChoOrig_XSizeConstrainedWindow_UpdateMeasure = XSizeConstrainedWindow.UpdateMeasure
+	AddToOrigs("XSizeConstrainedWindow.UpdateMeasure")
 	function XSizeConstrainedWindow.UpdateMeasure(...)
 		if UserSettings.StopSelectionPanelResize then
 			return ChoOrig_XWindow_UpdateMeasure(...)
@@ -419,6 +463,7 @@ function OnMsg.ClassesGenerate()
 
 	-- allows you to build outside buildings inside and vice
 	local ChoOrig_CursorBuilding_GameInit = CursorBuilding.GameInit
+	AddToOrigs("CursorBuilding.GameInit")
 	function CursorBuilding:GameInit(...)
 		if self.template_obj then
 			if UserSettings.RemoveBuildingLimits then
@@ -434,6 +479,7 @@ function OnMsg.ClassesGenerate()
 
 	-- stupid supply pods don't want to play nice (override for custom_travel_time_mars/custom_travel_time_earth)
 	local ChoOrig_RocketExpedition_ExpeditionSleep = RocketExpedition.ExpeditionSleep
+	AddToOrigs("RocketExpedition.ExpeditionSleep")
 	function RocketExpedition:ExpeditionSleep(s_t, ...)
 		if UserSettings.TravelTimeEarthMars then
 			s_t = g_Consts.TravelTimeEarthMars
@@ -442,6 +488,7 @@ function OnMsg.ClassesGenerate()
 	end
 
 	local ChoOrig_RocketBase_FlyToEarth = RocketBase.FlyToEarth
+	AddToOrigs("RocketBase.FlyToEarth")
 	function RocketBase:FlyToEarth(flight_time, ...)
 		if UserSettings.TravelTimeMarsEarth then
 			flight_time = g_Consts.TravelTimeMarsEarth
@@ -450,6 +497,7 @@ function OnMsg.ClassesGenerate()
 	end
 
 	local ChoOrig_RocketBase_FlyToMars = RocketBase.FlyToMars
+	AddToOrigs("RocketBase.FlyToMars")
 	function RocketBase:FlyToMars(cargo, cost, flight_time, ...)
 		if UserSettings.TravelTimeEarthMars then
 			flight_time = g_Consts.TravelTimeEarthMars
@@ -459,12 +507,14 @@ function OnMsg.ClassesGenerate()
 
 	-- no need for fuel to launch rocket
 	local ChoOrig_RocketBase_HasEnoughFuelToLaunch = RocketBase.HasEnoughFuelToLaunch
+	AddToOrigs("RocketBase.HasEnoughFuelToLaunch")
 	function RocketBase.HasEnoughFuelToLaunch(...)
 		return UserSettings.RocketsIgnoreFuel or ChoOrig_RocketBase_HasEnoughFuelToLaunch(...)
 	end
 
 	-- UI transparency cheats menu
 	local ChoOrig_XShortcutsHost_SetVisible = XShortcutsHost.SetVisible
+	AddToOrigs("XShortcutsHost.SetVisible")
 	function XShortcutsHost:SetVisible(...)
 		SetDlgTrans(self)
 		return ChoOrig_XShortcutsHost_SetVisible(self, ...)
@@ -472,6 +522,7 @@ function OnMsg.ClassesGenerate()
 
 	-- larger trib/subsurfheater radius
 	local ChoOrig_UIRangeBuilding_SetUIRange = UIRangeBuilding.SetUIRange
+	AddToOrigs("UIRangeBuilding.SetUIRange")
 	function UIRangeBuilding:SetUIRange(radius, ...)
 		local bs = UserSettings.BuildingSettings[self.template_name]
 		if bs and bs.uirange and not self:IsKindOf("TriboelectricScrubber") then
@@ -482,6 +533,7 @@ function OnMsg.ClassesGenerate()
 
 	-- override any performance changes if needed
 	local ChoOrig_Workplace_GetWorkshiftPerformance = Workplace.GetWorkshiftPerformance
+	AddToOrigs("Workplace.GetWorkshiftPerformance")
 	function Workplace:GetWorkshiftPerformance(...)
 		local set = UserSettings.BuildingSettings[self.template_name]
 		return set and set.performance_notauto or ChoOrig_Workplace_GetWorkshiftPerformance(self, ...)
@@ -489,6 +541,7 @@ function OnMsg.ClassesGenerate()
 
 	-- block certain traits from workplaces
 	local ChoOrig_Workplace_AddWorker = Workplace.AddWorker
+	AddToOrigs("Workplace.AddWorker")
 	function Workplace:AddWorker(worker, shift, ...)
 		local bs = UserSettings.BuildingSettings[self.template_name]
 		-- check that the tables contain at least one trait
@@ -533,16 +586,19 @@ function OnMsg.ClassesGenerate()
 
 		-- set amount of dust applied
 		local ChoOrig_BuildingVisualDustComponent_SetDustVisuals = BuildingVisualDustComponent.SetDustVisuals
+		AddToOrigs("BuildingVisualDustComponent.SetDustVisuals")
 		function BuildingVisualDustComponent:SetDustVisuals(dust, ...)
 			return ChangeDust(false, self, dust, ChoOrig_BuildingVisualDustComponent_SetDustVisuals, ...)
 		end
 		--
 		local ChoOrig_Building_SetDustVisuals = Building.SetDustVisuals
+		AddToOrigs("Building.SetDustVisuals")
 		function Building:SetDustVisuals(dust, ...)
 			return ChangeDust(false, self, dust, ChoOrig_Building_SetDustVisuals, ...)
 		end
 		--
 		local ChoOrig_DustGridElement_AddDust = DustGridElement.AddDust
+		AddToOrigs("DustGridElement.AddDust")
 		function DustGridElement:AddDust(dust, ...)
 			return ChangeDust(true, self, dust, ChoOrig_DustGridElement_AddDust, ...)
 		end
@@ -550,6 +606,7 @@ function OnMsg.ClassesGenerate()
 
 	-- change dist we can charge from cables
 	local ChoOrig_BaseRover_GetCableNearby = BaseRover.GetCableNearby
+	AddToOrigs("BaseRover.GetCableNearby")
 	function BaseRover:GetCableNearby(rad, ...)
 		local new_rad = UserSettings.RCChargeDist
 		if new_rad then
@@ -560,6 +617,7 @@ function OnMsg.ClassesGenerate()
 
 	-- add my cheats to the cheats pane
 	local ChoOrig_InfopanelObj_CreateCheatActions = InfopanelObj.CreateCheatActions
+	AddToOrigs("InfopanelObj.CreateCheatActions")
 	function InfopanelObj:CreateCheatActions(win, ...)
 		-- fire orig func to build cheats
 		if ChoOrig_InfopanelObj_CreateCheatActions(self, win, ...) then
@@ -600,6 +658,7 @@ function OnMsg.ClassesGenerate()
 		local margin = box(10, 0, 0, 0)
 
 		local ChoOrig_XMenuEntry_SetShortcut = XMenuEntry.SetShortcut
+		AddToOrigs("XMenuEntry.SetShortcut")
 		function XMenuEntry:SetShortcut(...)
 
 			if self.Icon == "CommonAssets/UI/Menu/folder.tga" then
@@ -682,6 +741,7 @@ function OnMsg.ClassesGenerate()
 
 	-- this one is easier than XPopupMenu, since it keeps a ref to the action (devs were kind enough to add a single line of "button.action = action")
 	local ChoOrig_XToolBar_RebuildActions = XToolBar.RebuildActions
+	AddToOrigs("XToolBar.RebuildActions")
 	function XToolBar:RebuildActions(...)
 		ChoOrig_XToolBar_RebuildActions(self, ...)
 		-- we only care for the cheats menu toolbar tooltips thanks
@@ -723,9 +783,15 @@ function OnMsg.ClassesGenerate()
 			[283142739680--[[Game]]] = "HideGameMenu",
 			[1000113--[[Debug]]] = "HideDebugMenu",
 			[487939677892--[[Help]]] = "HideHelpMenu",
+			Cheats = "HideCheatsMenu",
+			ECM = "HideECMMenu",
+			Game = "HideGameMenu",
+			Debug = "HideDebugMenu",
+			Help = "HideHelpMenu",
 		}
 
 		local ChoOrig_XMenuBar_RebuildActions = XMenuBar.RebuildActions
+		AddToOrigs("XMenuBar.RebuildActions")
 		function XMenuBar:RebuildActions(...)
 
 			ChoOrig_XMenuBar_RebuildActions(self, ...)
@@ -743,7 +809,8 @@ function OnMsg.ClassesGenerate()
 
 			for i = 1, #self do
 				local entry = self[i]
-				SetVis(entry, options, lookup_id[TGetID(entry.Text)])
+
+				SetVis(entry, options, lookup_id[entry.Text])
 			end
 		end
 	end -- do
@@ -754,6 +821,7 @@ function OnMsg.ClassesPostprocess()
 
 	-- align popups to rightside when using vertical cheat menu
 	local ChoOrig_XMenuBar_PopupAction = XMenuBar.PopupAction
+	AddToOrigs("XMenuBar.PopupAction")
 	function XMenuBar:PopupAction(action_id, ...)
 		if not ChoGGi.UserSettings.VerticalCheatMenu then
 			return ChoOrig_XMenuBar_PopupAction(self, action_id, ...)
@@ -769,6 +837,7 @@ function OnMsg.ClassesPostprocess()
 	end
 
 	local ChoOrig_SpaceElevator_DroneUnloadResource = SpaceElevator.DroneUnloadResource
+	AddToOrigs("SpaceElevator.DroneUnloadResource")
 	function SpaceElevator:DroneUnloadResource(...)
 		local export_when = ChoGGi.ComFuncs.DotPathToObject("ChoGGi.UserSettings.BuildingSettings.SpaceElevator.export_when_this_amount")
 		local amount = self.max_export_storage - self.export_request:GetActualAmount()
@@ -782,6 +851,7 @@ function OnMsg.ClassesPostprocess()
 	end
 
 	local ChoOrig_SpaceElevator_ToggleAllowExport = SpaceElevator.ToggleAllowExport
+	AddToOrigs("SpaceElevator.ToggleAllowExport")
 	function SpaceElevator:ToggleAllowExport(...)
 		ChoOrig_SpaceElevator_ToggleAllowExport(self, ...)
 		if self.allow_export and UserSettings.SpaceElevatorToggleInstantExport then
@@ -794,6 +864,7 @@ function OnMsg.ClassesPostprocess()
 
 	-- unbreakable cables/pipes
 	local ChoOrig_SupplyGridFragment_IsBreakable = SupplyGridFragment.IsBreakable
+	AddToOrigs("SupplyGridFragment.IsBreakable")
 	function SupplyGridFragment.IsBreakable(...)
 		if UserSettings.CablesAndPipesNoBreak then
 			return false
@@ -802,6 +873,7 @@ function OnMsg.ClassesPostprocess()
 	end
 	--
 	local ChoOrig_BreakableSupplyGridElement_CanBreak = BreakableSupplyGridElement.CanBreak
+	AddToOrigs("BreakableSupplyGridElement.CanBreak")
 	function BreakableSupplyGridElement.CanBreak(...)
 		if UserSettings.CablesAndPipesNoBreak then
 			return false
@@ -811,6 +883,7 @@ function OnMsg.ClassesPostprocess()
 
 	-- no more pulsating pin motion
 	local ChoOrig_XBlinkingButtonWithRMB_SetBlinking = XBlinkingButtonWithRMB.SetBlinking
+	AddToOrigs("XBlinkingButtonWithRMB.SetBlinking")
 	function XBlinkingButtonWithRMB:SetBlinking(...)
 		if UserSettings.DisablePulsatingPinsMotion then
 			self.blinking = false
@@ -841,12 +914,14 @@ function OnMsg.ClassesPostprocess()
 
 		-- no more stuck focus on ECM textboxes/lists
 		local ChoOrig_XDesktop_MouseEvent = XDesktop.MouseEvent
+		AddToOrigs("XDesktop.MouseEvent")
 		function XDesktop:MouseEvent(event, ...)
 			return ResetFocus(ChoOrig_XDesktop_MouseEvent, self, event, ...)
 		end
 
 		-- make sure focus isn't on my dialogs if gamepad is in play
 		local ChoOrig_XDesktop_XEvent = XDesktop.XEvent
+		AddToOrigs("XDesktop.XEvent")
 		function XDesktop:XEvent(event, ...)
 			return ResetFocus(ChoOrig_XDesktop_XEvent, self, event, ...)
 		end
@@ -854,6 +929,7 @@ function OnMsg.ClassesPostprocess()
 
 	-- removes earthsick effect
 	local ChoOrig_Colonist_ChangeComfort = Colonist.ChangeComfort
+	AddToOrigs("Colonist.ChangeComfort")
 	function Colonist:ChangeComfort(...)
 		ChoOrig_Colonist_ChangeComfort(self, ...)
 		if UserSettings.NoMoreEarthsick and self.status_effects.StatusEffect_Earthsick then
@@ -863,6 +939,7 @@ function OnMsg.ClassesPostprocess()
 
 	-- make sure heater keeps the powerless setting
 	local ChoOrig_SubsurfaceHeater_UpdateElectricityConsumption = SubsurfaceHeater.UpdateElectricityConsumption
+	AddToOrigs("SubsurfaceHeater.UpdateElectricityConsumption")
 	function SubsurfaceHeater:UpdateElectricityConsumption(...)
 		ChoOrig_SubsurfaceHeater_UpdateElectricityConsumption(self, ...)
 		if self.ChoGGi_mod_electricity_consumption then
@@ -872,6 +949,7 @@ function OnMsg.ClassesPostprocess()
 
 	-- same for tribby
 	local ChoOrig_TriboelectricScrubber_OnPostChangeRange = TriboelectricScrubber.OnPostChangeRange
+	AddToOrigs("TriboelectricScrubber.OnPostChangeRange")
 	function TriboelectricScrubber:OnPostChangeRange(...)
 		ChoOrig_TriboelectricScrubber_OnPostChangeRange(self, ...)
 		if self.ChoGGi_mod_electricity_consumption then
@@ -881,6 +959,7 @@ function OnMsg.ClassesPostprocess()
 
 	-- remove idiot trait from uni grads (hah!)
 	local ChoOrig_MartianUniversity_OnTrainingCompleted = MartianUniversity.OnTrainingCompleted
+	AddToOrigs("MartianUniversity.OnTrainingCompleted")
 	function MartianUniversity:OnTrainingCompleted(unit, ...)
 		if UserSettings.UniversityGradRemoveIdiotTrait then
 			unit:RemoveTrait("Idiot")
@@ -923,11 +1002,13 @@ function OnMsg.ClassesPostprocess()
 		end
 
 		local ChoOrig_SA_WaitTime_StopWait = SA_WaitTime.StopWait
+		AddToOrigs("SA_WaitTime.StopWait")
 		function SA_WaitTime:StopWait(...)
 			return SkipMystStep(self, ChoOrig_SA_WaitTime_StopWait, ...)
 		end
 		--
 		local ChoOrig_SA_WaitMarsTime_StopWait = SA_WaitMarsTime.StopWait
+		AddToOrigs("SA_WaitMarsTime.StopWait")
 		function SA_WaitMarsTime:StopWait(...)
 			return SkipMystStep(self, ChoOrig_SA_WaitMarsTime_StopWait, ...)
 		end
@@ -935,6 +1016,7 @@ function OnMsg.ClassesPostprocess()
 
 	-- keep prod at saved values for grid producers (air/water/elec)
 	local ChoOrig_SupplyGridElement_SetProduction = SupplyGridElement.SetProduction
+	AddToOrigs("SupplyGridElement.SetProduction")
 	function SupplyGridElement:SetProduction(new_production, new_throttled_production, update, ...)
 		local amount = UserSettings.BuildingSettings[self.building.template_name]
 		if amount and amount.production then
@@ -954,6 +1036,7 @@ function OnMsg.ClassesPostprocess()
 
 	-- and for regular producers (factories/extractors)
 	local ChoOrig_SingleResourceProducer_Produce = SingleResourceProducer.Produce
+	AddToOrigs("SingleResourceProducer.Produce")
 	function SingleResourceProducer:Produce(amount_to_produce, ...)
 		local amount = UserSettings.BuildingSettings[self.parent.template_name]
 		if amount and amount.production then
@@ -982,11 +1065,13 @@ function OnMsg.ClassesPostprocess()
 		end
 
 		local ChoOrig_RCRover_SetWorkRadius = RCRover.SetWorkRadius
+		AddToOrigs("RCRover.SetWorkRadius")
 		function RCRover:SetWorkRadius(radius, ...)
 			SetHexRadius(ChoOrig_RCRover_SetWorkRadius, "RCRoverMaxRadius", self, radius, ...)
 		end
 		--
 		local ChoOrig_DroneHub_SetWorkRadius = DroneHub.SetWorkRadius
+		AddToOrigs("DroneHub.SetWorkRadius")
 		function DroneHub:SetWorkRadius(radius, ...)
 			SetHexRadius(ChoOrig_DroneHub_SetWorkRadius, "CommandCenterMaxRadius", self, radius, ...)
 		end
@@ -994,6 +1079,7 @@ function OnMsg.ClassesPostprocess()
 
 	-- toggle trans on mouseover
 	local ChoOrig_XWindow_OnMouseEnter = XWindow.OnMouseEnter
+	AddToOrigs("XWindow.OnMouseEnter")
 	function XWindow:OnMouseEnter(...)
 		if UserSettings.TransparencyToggle then
 			self:SetTransparency(0)
@@ -1002,6 +1088,7 @@ function OnMsg.ClassesPostprocess()
 	end
 	--
 	local ChoOrig_XWindow_OnMouseLeft = XWindow.OnMouseLeft
+	AddToOrigs("XWindow.OnMouseLeft")
 	function XWindow:OnMouseLeft(...)
 		if UserSettings.TransparencyToggle then
 			SetDlgTrans(self)
@@ -1017,6 +1104,7 @@ function OnMsg.ClassesPostprocess()
 		local UnbuildableZ = buildUnbuildableZ()
 
 		local ChoOrig_ConstructionController_UpdateCursor = ConstructionController.UpdateCursor
+		AddToOrigs("ConstructionController.UpdateCursor")
 		function ConstructionController:UpdateCursor(pos, force, ...)
 			if self.is_template and IsValid(self.cursor_obj)
 				and self.template_obj.dome_spot == "Spire"
@@ -1113,6 +1201,7 @@ function OnMsg.ClassesPostprocess()
 
 	-- tweak console when it's "opened"
 	local ChoOrig_Console_Show = Console.Show
+	AddToOrigs("Console.Show")
 	function Console:Show(show, ...)
 		ChoOrig_Console_Show(self, show, ...)
 		if show then
@@ -1152,6 +1241,7 @@ function OnMsg.ClassesPostprocess()
 		}
 		-- skip quit from being added to console history to prevent annoyances
 		local ChoOrig_Console_AddHistory = Console.AddHistory
+		AddToOrigs("Console.AddHistory")
 		function Console:AddHistory(text, ...)
 			if skip_cmds[text] or text:sub(1, 5) == "quit(" then
 				return
@@ -1164,6 +1254,7 @@ function OnMsg.ClassesPostprocess()
 	-- I could do a thread and wait till the key isn't pressed, but it's slower
 	-- this does block user from typing in `, but eh
 	local ChoOrig_Console_TextChanged = Console.TextChanged
+	AddToOrigs("Console.TextChanged")
 	function Console:TextChanged(...)
 		ChoOrig_Console_TextChanged(self, ...)
 		local text = self.idEdit:GetText()
@@ -1182,11 +1273,13 @@ function OnMsg.ClassesPostprocess()
 		end
 
 		local ChoOrig_Console_HistoryDown = Console.HistoryDown
+		AddToOrigs("Console.HistoryDown")
 		function Console:HistoryDown(...)
 			HistoryEnd(ChoOrig_Console_HistoryDown, self, ...)
 		end
 		--
 		local ChoOrig_Console_HistoryUp = Console.HistoryUp
+		AddToOrigs("Console.HistoryUp")
 		function Console:HistoryUp(...)
 			HistoryEnd(ChoOrig_Console_HistoryUp, self, ...)
 		end
@@ -1197,6 +1290,7 @@ function OnMsg.ClassesPostprocess()
 		local tonumber = tonumber
 
 		local ChoOrig_RequiresMaintenance_AddDust = RequiresMaintenance.AddDust
+		AddToOrigs("RequiresMaintenance.AddDust")
 		function RequiresMaintenance:AddDust(amount, ...)
 			-- maybe something was sending a "number" instead of number?
 			amount = tonumber(amount)
@@ -1212,6 +1306,7 @@ function OnMsg.ClassesPostprocess()
 		local UpdateConstructionStatuses = ConstructionController.UpdateConstructionStatuses
 
 		local ChoOrig_TunnelConstructionController_UpdateConstructionStatuses = TunnelConstructionController.UpdateConstructionStatuses
+		AddToOrigs("TunnelConstructionController.UpdateConstructionStatuses")
 		function TunnelConstructionController:UpdateConstructionStatuses(...)
 			local skip
 			if UserSettings.RemoveBuildingLimits then
@@ -1254,6 +1349,7 @@ function OnMsg.ClassesPostprocess()
 
 		-- show scroll on hover
 		local ChoOrig_InfopanelDlg_OnMouseEnter = InfopanelDlg.OnMouseEnter
+		AddToOrigs("InfopanelDlg.OnMouseEnter")
 		function InfopanelDlg:OnMouseEnter(...)
 			-- show scrollbar
 			if UserSettings.ScrollSelectionPanel and infopanel_list[self.XTemplate]
@@ -1266,6 +1362,7 @@ function OnMsg.ClassesPostprocess()
 		end
 
 		local ChoOrig_InfopanelDlg_OnMouseLeft = InfopanelDlg.OnMouseLeft
+		AddToOrigs("InfopanelDlg.OnMouseLeft")
 		function InfopanelDlg:OnMouseLeft(...)
 			-- hide scrollbar
 			if UserSettings.ScrollSelectionPanel and infopanel_list[self.XTemplate]
@@ -1505,6 +1602,7 @@ function OnMsg.ClassesPostprocess()
 
 		-- the actual function
 		local ChoOrig_InfopanelDlg_Open = InfopanelDlg.Open
+		AddToOrigs("InfopanelDlg.Open")
 		function InfopanelDlg:Open(...)
 			CreateRealTimeThread(function()
 				WaitMsg("OnRender")
@@ -1650,6 +1748,7 @@ end]]
 
 		-- ReadHistory fires from :Show(), if it isn't loaded before you :Exec() then goodbye history
 		local ChoOrig_Console_Exec = Console.Exec
+		AddToOrigs("Console.Exec")
 		function Console:Exec(text, hide_text, ...)
 			if not self.history_queue or #self.history_queue == 0 then
 				self:ReadHistory()